refactor(leumi): split forex helpers into leumi-forex.ts, add unit tests
Addresses @baruchiro's review on PR #1149: moves the Leumi scraper into
src/scrapers/leumi/, extracts the foreign-currency parsing/fetching logic
(mapForeignCurrency, parseForeignAmount, mapForeignTransaction, etc.) into
its own leumi-forex.ts, and adds leumi-forex.test.ts covering the pure
helpers with real observed label/row examples as a TDD base for future
Leumi forex changes.
